+++ CHANGELOG.md
@@ -4,11 +4,19 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2019-12-03
+
+### Adaptive Performance Simulator and Extension
+- Add Adaptive Performance Simulator and Device Simulator extension to simulate Adaptive Performance events in the editor using the device simulator package. 
+
+### Bump Min supported Version
+- Adaptive Performance Device Simulator extension requires the Device Simulator package which uses features of Unity 2019.3 and therefore we update Adaptive Performance 2.0.0 min support to 2019.3. Please use Adaptive Performance 1.x for earlier versions supporting 2018.3+.  
+
 ## [1.1.0] - 2019-11-07
 
 ### Fix warnings and logging
 - Fixed initial version log to represent the correct version (subsystem)
-- Fixed non reachable code warning. 
+- Fixed non reachable code warning 
 
 ## [1.0.1] - 2019-08-29